Make pipeline work
