+++ scanpipe/pipes/strings.py
@@ -20,7 +20,7 @@
 # ScanCode.io is a free software code scanning tool from nexB Inc. and others.
 # Visit https://github.com/nexB/scancode.io for support and download.
 
-from scanpipe.pipes import LoopProgress
+from pipeline import LoopProgress
 
 
 class XgettextNotFound(Exception):
